requirements.txt
django==3.1.4
pylint==2.4.4
psycopg2-binary==2.8.4
python-decouple==3.3
django-bootstrap4==1.1.1
requests==2.23.0
gunicorn==20.0.4